Exclude unused material from pre CLAR rep orders
Because we do not have a CLAR fee scheme for LGFS
we need to compare the rep order date and the clar
release date to determine whether the fee is
eligible.

Note that if there is no rep order date we default to
include the unused materials - this is inline with
validations (and which is necessary because api
submissions may not have rep orders yet).