Add sentry to api sandbox
we have an old sentry account for the api sandbox
but the sandbox itself is not configured to send
errors to it. This was probably missed as part
of k8s migration so am adding back in now
after having created a new sentry account for it
on the new sentry box.