Set eager_load to true for test environment
Setting to true to avoid LoadError problems for which i can find
no other solution. Specifically the cucumber test suite encounters
```
<LoadError: Unable to autoload constant ExternalUsers::Fees::PricesController, ...
<LoadError: Unable to autoload constant CCR::CaseTypeAdapter, ...
<LoadError: Unable to autoload constant CCLF::CaseTypeAdapter, ...
```
...if not eagerloaded.

This requires env variables to be defined in `.env.test` for local
runs of the test suite. These can be copied from `.env.sample` but
can probably be refined to a bare minimum.

Also, need to look at excluding some lib folders and ensuring we are
using Zeitwork mode really.

Have had to prevent eager loading of demo data (which has files in the lib
folder that probably should not be there).