Change sample user scripts to classes
So they are only loaded by explicit calling.

Beacuse these files exist in the lib folder
any eager loading will run them. Since they
are scripts this means an eager load will
run them. This is not desirable as we only
need to run these seeds via rake tasks or
on request, such as for the smoke test.