Add kubernetes config for dump job and amend job script
The dump job will run a rake task that dumps the
database on a given namespace/environment.

The job script is modified to wait for the job pod to
start and then tail the logs so you can view progress.

Note: there is a timeout on logs --follow which, i believe,
is an IdleTimeout. It appears to be about 5 minutes. It may be
possible to prevent this by outputting any log entry every x seconds.