Add kubernetes anonymised db dump task and reorganise
The new tasks are for use in dumping and uploading
dump files to s3, as well as retrieving them from s3
to the local machine.

The basic flow is
  - rake task db:dump:run_job, uses a script to run
    the k8s dump job (script tails job log)
  - job executes db:dump:anonymised rake task
    which creates anonoymised dump file, compresses it and
    uploads it to s3 and then all lists dump files to stdout.
  - rake task db:dump:copy_s3_dump copies file from s3 to local
    machine (requires git-crypted secrets access).
  - rake task db:restore is used to restore it to local
    machine

There was a lot of legacy unusable code that has also
been replaced or removed. Shared reusable code has
been moved to `rake_helpers` directory and required
/included where required.