Rename and add MI date fields
Case workers are confused by the name "Submitted at".

They need to see either the last_submitted_at date or
the original_submission_date, but we are not sure what is
most applicable for their purposes.

So, we are adding the two new dates and they can then decide
what they want.

"Submitted at" column in MI is the date the claim transitioned
to a state of submitted, awaitiing_written_reasons or
redetermination.