Ensure messages are created before test
The tests only work by a quirk of RSpec. `claim.unread_messages_for` returns an unevaluated
ActiveRecord relation that, if it were evaluated immediately, would return an empty list.
The `match_array` matcher, however, evaluates the array `[message_1]` first, ensuring that
`message_1` is added to the database before the relation is evaluated.