CBO-1632 Hide typeahead selector
Before change, if type selector (`select`)  was removed, the test would still pass,
this is because the element was visually hidden and the spec could still interact with it.

Now, the selector will be entirely hidden, therefore, spec will fail.

Note: The typeahead plugin copies the select's classes to a newly created input,
therefore our css needs to specifically target the select field, thus we disable stylelint on the line.