Get skipped test passing
This has been skipped for a long time
and was going to delete but unclear
wether the functionality is still required.

Assuming it is and therefore it should be
tested.

Cannot say i can see anyway this scenario
will ever come up though, as the comment
on the test says.