Fix edge case behaviour
There is a test that tests for what would now
be a situation which is to my mind currently
impossible. Nameley that a user only has one
claim type available to them and that this
claim type is not a known, valid type.

The test and code, in such situations
would redirect the user back to there
home page.