Anonymise entire filename
Previously the extension (text after last dot)
was not anonymised. Hwoever, some files with extremely
long names are being uploaded and the names truncated.
this can result in an unanomised part of the file name.

This solution renames the file and relies on the
mime content_type to then reattach the correct
extension.