Add checksum to Paperclip attachments
A new mixin, CheckSummable, is added to Message, Document and StatsReport so
that a checksum is calculated for all newly created assets and stored in the
database alongside the Paperclip fields. This checksum is required by Active
Storage and will be copied over to the new database tables when the assets are
migrated.