Update Stats::StatsReport to use Active Storage
Stats reports are now stored using Active Storage instead of Paperclip.
Additionally, report downloads are done using a redirect to an expiring url
in S3 instead of pulling the file to the server and then passing on to the
user. A custom redirect action:

* `CaseWorkers::Admin::ManagementInformationController#download`

instead of the built in helpers `rails_blob_path` or `rails_blob_url` as these
provide a permanent (i.e., not expiring) url for accessing the asset without
authentication.