Handle "Specified key does not exist errors" for dummy file creation
When creating a dummy file on s3 paperclip writes
to STDOUT a warning that the "specified key does not exist".

This is because while the paperclip meta
data exists from the DB restore the actual
file does not YET exist. This is just noise
in the context of dummy file creation so
we can ignore.