Refactor to fix Metrics/AbcSize Rubocop offense
After fixing `ErrorPresenter.generate_messages` to use Active Model errors
correctly for Rails 6.1 it triggered a Metrics/AbcSize Rubocop offense. I have
made a small refactor that clears this but more could be done.

When iterating over `@errors` there is now just one loop variable representing
each error instead of the field name and message, as there was before. This is
now passed to `populate_messages` as a single argument instead of three,
including an instance of `ErrorMessageTranslator`.

Further work could be done to move the logic inside `populate_messages` into
`ErrorMessageTranslator` and update `ErrorDetails` so that its initializer
can take an instance of `ErrorMessageTranslator` as an argument. This is
(probably) outside of the scope of this PR (upgrading to Rails 6.1).