Prevent overwritting of existing files
Creation of dummy files for evidence documents
failed cease with an unknown error after 20
minutes and having created some 6 thousdand.

Not sure of the cause (could be a timeout issue
on runniing the command unattended) but
don't want to have to recreate dummy files again
so putting a check if doc exists. This adds overhead
however :(