Add ActionController::InvalidAuthenticityToken handler
This error would be raised when the secret_key_base
is rotated without an old_secret_key_base to fallback on
(see prevous commits to handle seamless cookie rotation).

Currently this error would be handled as a general exception
and cause a 500 rather than redirecting the user to login again.
A more pleasant user experience in such situations would be to
,at least, redirect them to login again.

NOTE: devise may make this redundant, need to test