Reduce nesting of blocks
Rubocop flags if nesting in RSpec tests is too deep (RSpec/NestedGroups). To
reduce this the context around each instance of the advocate categories shared
examples is removed. An extra context is added in each of these shared examples
blocks so that the information is not lost when tests fail.