Fetch attachment path from Active Storage
This allows a rollback from Active Storage to Paperclip.

If an attachment is added using Active Storage the path that is used by
Paperclip to fetch it will fail. The solution is to check to see if an
Active Storage record exists with a key not containing a '/' character (which
would indicate a record previously migrated from Paperclip) then the path
used is calculated by `ActiveStorage::Blob.service.path_for` instead of
Paperclip's pattern.

To enable this to work the stats reports, messages and documents models, when
they are switched to Active Storage, will need to keep populating the Paperclip
fields.