Handle api/custom_validations/date_format keyword parameter deprecation
Fix the deprecation, that will break in ruby 3.0
``
app/interfaces/api/custom_validations/date_format.rb:24: warning: Using the last argument as keyword parameters is deprecated; maybe ** should be added to the call
```

Rather than fixing all calls to this method
from specs and elsewhere it is cleaner/less
verbose to avoid the problem by changing the
underlying code causing it. This the raising
of an error using the simpler verion of passing
a class as an argument with additional arguments
appended. Instead by initializing an instance
of the class and passing the arguments to THAT
avoids the problem.

see the grape validation error class concerned.
```
gems/grape-1.5.3/lib/grape/exceptions/validation.rb
```