Fix call to govuk_component
This call previously worked as kwargs were correctly
mapped to **tag_options in ruby 2.6.5. Now, because the
second argument is passed to the second parameter regardless
the `class: 'fx-duplicate-expense'` was being intepreted by
url_for().

So
```
= govuk_link_button_secondary(t('.duplicate'), class: 'fx-duplicate-expense')
```
ends up calling:
```
url_for({class: 'fx-duplicate-expense'`})
=> /path/to/page?class=fx-duplicate-expense
```

Passing `nil` as the second argument is sufficient
to fix this problem as url_for fallsback to using current path
in controllers and views.
```
url_for(nil)
=> /path/to/page
```