Extend and document intended use of `govuk_link_button`
A bump to ruby 2.7.1 caused a breakage with
no tests covering it. A feature test has
been added to cover this previously from the app level.

Some unit specs should also be added cover and document the
intended happy path for this helper and those
that depend on it - the `govuk_link_button_secondary`
and `govuk_link_button_warning`.