Bump mimemagic to 3.7.0
Version 3.6.0 has been yanked and 3.7.0 has been released with the MIT licence
again following the removal of freedesktop.org.xml.

This file exists at /usr/share/mime/packages/freedesktop.org.xml in our docker
images and so this might be enough. If not it will be necessary to set the
`FREEDESKTOP_MIME_TYPES_PATH` environment variable.