Treat zero `calculate_distance` as valid
Currently with the google maps API broken
the request is returning nil which is converted
to zero. To allow for google maps API failures
without raising a validation error to the user
we must allow zeros to not be invalid.

This allows the user to enter a manually travel
distance which case workers can then vet manually.