Amend calculated_distance minimum validation
A simpler, and IMO clearer, way to loosen
the validation.

Rather than ensure a non-zero value before attempting
to validate we can just set the minimum to be 0.0. A
failed API call, returning 0, would then be considered valid,
while a negative decimal would be invalid. In addition this
validates upper limit by default (of 200_000).