Correct advocate_supplementary_claim_submit feature test
The uplift fee was wrong but due to an odd behaviour
related to ordering in postgres 9.6 and the JS for
calculating uplift fees this was specing incorrect
behaviour.

The misc fee types are ordered by their description on the page.
Local postgres 9.6.21 orders strings with parenthesis slightly
differently:
```
 [[40, "Confiscation hearings (half day uplift)"],
  [38, "Confiscation hearings (half day)"],...
 ]
```
which does not appear to be the case on any of the environments though
as they are ordering as expected:
````
 [[38, "Confiscation hearings (half day)"],
  [40, "Confiscation hearings (half day uplift)"],...
 ]
```
The order of the fees on the page then has an impact on fee calculation for uplifts,
via the JS in some way, such that if an uplift is above its non-uplift fee equivalent
it will not be recalculated.

Ancedotally, postgres 13.1, appears to implicitly be handling parentheses ordering as
we would expect and thereby calculating the correct amount. The feature test is testing
the wrong amount.