Enforce expected order of misc fee types by default
Because the order of the fees appears to impact fee calculation
of uplifts.

Explanation:
The misc fee types are ordered by their description on the page.
Local postgres 9.6.21 orders strings with parentheses slightly
differently:
```
 [[40, "Confiscation hearings (half day uplift)"],
  [38, "Confiscation hearings (half day)"],...
 ]
```

this does not appear to be the case on any of the environments
though as they are ordering as expected.
```
 [[38, "Confiscation hearings (half day)"],
  [40, "Confiscation hearings (half day uplift)"],...
 ]
```

The order of the fees on the page then has an impact on fee
calculation for uplifts, via the JS in some way, such that
if an uplift is above its non-uplift fee equivalent it will
not be recalculated.