Enforce expected order of misc fee types by default
Because of difference in ordering of strings generally
and fee type description specifically, depending on postgres
host (darwin/alpine/amazonlinux).

Explanation:
The misc fee types are ordered by their description on the page.
Local postgres 9.6.21/darwin orders strings with parentheses and
whitespace slightly differently:
```
 [[40, "Confiscation hearings (half day uplift)"],
  [38, "Confiscation hearings (half day)"],...
 ]
```
Alpine hosted postgres (circleci postgres alpine images) behave similarly.

this does not appear to be the case on any of the RDS backed (amazon linux)
environments though as they are ordering as expected. Debian hosted postgres
behaves similarly.
```
 [[38, "Confiscation hearings (half day)"],
  [40, "Confiscation hearings (half day uplift)"],...
 ]
```

The order of the fees on the page is WAS significant due to a javascript
bug, but this has been corrected in this branch/PR now.