Use virtual attribute to require T&Cs to be agreed
It has been agree that user creation should not require
agreeing terms_and_conditions unless they are signing
themselves up - self registration. Self registration
is only possible on the api-sandbox in any event.

When an admin for a provider creates a user in their
provider, or when a case worker admin creates an external
user or case worker on the sandbox (or anywhere else) they
should not need to agree terms and conditions.