Amend cronjobs to always use app-latest image
Jobs pods spun up by the archvive_stale cronjob
are not being kept around.

I wonder if it is because when a new image is
applied during deploy this causes old job pods
to be removed.

We want see these logs to check archiving (or vacuuming)
was successful or not, and details thereof.