Add on_home_page? helper and spec
Requirement is to display the message
only on the home page. This is problematic
since different user types have different
home pages, plus some have different home
pages for different roles (admin and "non-admin"
caseworkers). In addition hit the title link
of the app uses the "root" path '/'

This list appears to be cover them all and can
therefore be used to show/hide the downtime
banner only on the "home page".

Note: also moved downtime helper from claim helper
to application helper because this is is an application
level feature rather than claim level.