Add litigator "no indictment" popup feature test
This feature had been indvertently broken for while
due to suttle view/frontend changes that impacted it.

Jasmine unit tests had not picked this up as we need
an "integration" test to cover the functionality.

Extend to cover:
 - indictment checkbox ticked --> warning not displayed
 - advocate grad fee claim (Trial) --> warning displayed
 - advocate fixed fee claim (Appeal..) --> warning NOT displayed

NOTE:

Add brief sleep after checking checkbox

Have tried many different mechanisms to prevent
an error which results when:
 - popups are enabled
 - you check an evidence checkbox
 - you hit "Save and continue" button

This is the only mechanism which
consistently succeeds. It may be related to the NewClaim.js
code for displaying the popup.