Add alternative name for @caseworker-seed-requirements hook
This Before hook just adds vat seeds which is useful in other
scenarios not requiring full seed, not just for caseworkers.

Specifically when creating a claim using factories there
is a dependency on VAT rates #2:

Also, rename the OS variable used to store the hooks status
to match, for clarity.