Update allocation notification
The allocation page is the root page when logged in as a caseworker,
This page would then have 2 notification banners,
1) downtime banner
2) allocation notification banner

Upon allocation submit, `Modules.AllocationDataTable.js` would update the content
of both banners. This is an incorrect behaviour.

We add an ID to add specificity to target the allocation notification banner
instead of both banners.