Add unused materials notice
Providers are being underpaied for work they are entitled to claim for.
A notice is put on the miscelleneous fees page in an AGFS final, LGFS final
or LGFS transfer claim with a rep order dated on or after 17th September 2020
to indicate that it may be eligible for unusted materials fees.