Fix validation error preventing claim creation
Unclear why this error is happening now for this
one test when it should have been failing previously.

```
Failure/Error: claim.reload

ActiveRecord::RecordNotFound:
  Couldn't find Claim::LitigatorClaim without an ID
./spec/api/v2/cclf_claim_spec.rb:88:in `create_claim'
./spec/api/v2/cclf_claim_spec.rb:280:in `block (7 levels) in <top (required)>'
./spec/api/v2/cclf_claim_spec.rb:305:in `block (8 levels) in <top (required)>'
./spec/vcr_helper.rb:84:in `block (3 levels) in <top (required)>'
./spec/vcr_helper.rb:83:in `block (2 levels) in <top (required)>'
```

The fix is, however, to correct the validation error - which was
caused by not having a case type to which you can add an unused materials
fee.