Fix a regression relating to error summaries
Additional submodel errors were appearing on the
model instances and then the presenter was falling
back to displaying them in the summary.

These errors are duplicates of other errors already
being handled by the custom error handler and, in
addition, do not result in a working links to the
relevant (or any) field.

I have widdened the deletion to delete any error
with an attribute indicating it is a nested error.

This should be safe to do as such errors are copied
to the parent/base record in a specifci custom format
for handling by the custom error presenter.

I have sanity tested its impact on a final claim for fixed fees