Move eligible_misc_fee_types to Claim::BaseClaim
`eligible_misc_fee_types` is identical in all places where it is defined and
the only reason it is not in `Claim::BaseClaim` is because it does not apply
for some claims. However, `Claims::FetchEligibleMiscFeeTypes` can be changed
to deal with the exceptional cases (interim claims) in a sensible way by
returning an empty array, indicating that no misc fee types are eligible.

This is done so that it can be used to determine whether information messages
can be displayed to the user depening on whether certain fees can be claimed.