Use default url for form
The url `provider_management_provider_external_user_path(@provider, @external_user)`
that is given for the form is specifically for an update on an existing
external user but this form needs to apply to creation of new users.

`form_with model: [:provider_management, @provider, @external_user]` without a
`url` parameter correctly sets a default url based on the namespace
(`provider_management`) and models (`@provider` and `@external_user`) provided
and this handles a new or existing `@external_user`.