Add concern to handle govuk_design_system_form_builder error linking
Current validators use error attribute naming that
breaks linking of summary and field level errors when
using the govuk_design_system_form_builder.

This extracts the naming to its own method and
uses a composition approach to overide the
inherited behaviour from the `base_sub_model_validator`
for those sub model validators that exist on models
that are using the govuk_design_system_formbuilder
for their front-end forms.

This also makes the inherited version of the
`associated_error_attribute` method public so
we can test the inherited behaviour as well
as composed behaviour.

  Testing
  -------
  Add missing unit tests for
   - supplier_number_sub_model_validator_spec
   - defendant_sub_model_validator_spec
   - fee_sub_model_validator_spec

  Use shared examples to cover mixin behaviour and default (inherited)
  behaviour.

  Additional
  -------
  In addition this removes the `suffix_error_fields?` method
  because it defaults to true and is only overridden in one
  validator which also returns true. Therefore is alwasy true.