Add the submitted claim report to the reports page
A new report is added to show a count of all submitted claims per week for the
past twelve weeks. The rake task to create the submitted claims report is
updated to use the same report generation instead of its own helper.

This new report has been created following the pattern of the provisional
assessment report and rejections refusals report, which both work in the same
way. Specifically, the `Stats::ProvisionalAssessmentReportGenerator` and
`Stats::RejectionsRefusalsReportGenerator` classes are services that are
identical except for the class that is used for generating the data. With this
in mind I have pulled out the common functionality of these into
`Stats::ReportGenerator::Base` and built the new service
`Stats::ReportGenerator::SubmittedClaims` on top of this.

After this (in separate commits) the provisional assessment and rejectsions
refusals reports can be refactored to also be built on top of
`Stats::ReportGenerator::Base`.