Consolidate report generators
The three report generators:

* `app/services/stats/report_generator/provisional_assessment.rb`
* `app/services/stats/report_generator/rejections_refusals.rb`
* `app/services/stats/report_generator/submitted_claims.rb`

are all based off `app/services/stats/report_generator/base.rb` and only differ
in a single parameter. The `base.rb` can be generalised to take a the parameter
as an argument to remove an unnecessary layer.

The management information generator does not follow the same pattern although
it could be changed to avoid requiring an exception in `generate_new_report`
in `Stats::StatsReportGenerator`.

The tests for the three report generators that have been consolidated have been
moved into `spec/services/stats/report_generator_spec.rb