Apply govuk-font-family to app
With the migration to the GOV.UK Frontend typography
and the removal of the deprecated typography stylesheet imports,
element and or components not yet migrated, use Times New Roman font.

This commit sets govuk-font-family-gds-transport on the <html> tag.

However, while the application is still running in compatibility mode
as we continue the migration, the app will use Arial font as the default.