Set default_protect_from_forgery to true
Prior to 5.2 the ApplicationController required

```
protect_from_forgery prepend: true, with: :exception
```

This is now added to ActionController::Base, but only if
default_protect_from_forgery is set to true. This becomes the default setting
when the 5.2 defaults are loaded. This commit prepares for this by setting it
explicitly and updating the controllers accordingly.