Remove ssl redirect exclusion for `/ping` and `/healthcheck`
Since the migration to kubernetes, at least, these
endpoints are SSL/TLS enabled so the readiness and liveness
probes should work.

Testing this first to see if it can successfully be deployed
by responding successfully to the readiness probe.

Next step will be to look at enabling:
```
hsts: { subdomains: true }
```