Remove unnecessary hidden field
The hidden 'scheme' field is already being created in the 'scheme_filters'
partial. When the `form_with_generates_ids` configuration option is set to
true the field created in the partial has an id added to it, resulting in an
accessibility fault due to two fields with the same id. The field created in
`_allocation.html.haml` is unnecessary and can be removed.