Adopt rails 6.1 defaults for active_job options
 - active_job.retry_jitter to apply random job delay on retr.
   harmless!
 - active_job.skip_after_callbacks_if_terminated
   we do not use `after_enqueue` or `after_perform` so not impact

This allows us to adopt config.load_defaults "6.1", but

NOTE: the config setting is not taking effect since it appears
that govuk_notify_rails is causing early rails loading and
causing it to be "lost".

Therefore to make the setting stick we need apply directly to
`ActiveJob::Base`. This needs fixing in the future.