Do not adopt default for `action_mailer.deliver_later_queue_name`
Continue using :mailers queue for mail processing to keep
its priority low - its quick enough!

Baseline default is :mailers which is what we currently
use. Rails 6.1 default of nil results in using active_job.default_queue_name,
which is the 'default' queue.

However:

NOTE: the config setting is not taking effect since it appears
that govuk_notify_rails is causing early rails loading and
causing it to be "lost".

Therefore to make the setting stick we need apply directly to
`ActionMailer::Base.deliver_later_queue_name`. This needs fixing
in the future.