Adopt rails 6.1 default for `legacy_connection_handling` (false)
Rails 6. Default of false enables use of new DB connection handling API.

 > For applications using multiple databases, this new API provides
   support for granular connection swapping.

Although we do not use multiple databases the new API should be
adopted going forward and allows us to us config.load_defaults "6.1"
in the future.

However:

NOTE: the config setting is not taking effect since it appears
that govuk_notify_rails is causing early rails loading and
causing it to be "lost".

Therefore to make the setting stick we need apply directly to
`ActiveRecord::Base.legacy_connection_handling`. This needs fixing
in the future.