Adopt rails 6.1 default for `active_record.has_many_inversing` (true)
Enable easier bi-directional Primay key to foreign key
relation navigation.

 > enables setting the inverse record when traversing belongs_to
   to has_many associations.

See [this article](https://www.bigbinary.com/blog/rails-6-1-adds-support-for-belongs_to-to-has_many-inversing) for details

Since we use `inverse_of` of a lot in CCCD this should be beneficial
and mean we could remove `some_relation.reload`. Its unclear if
we need to add `inverse_of` to get this behaviour or not.