Adopt rails 6.1 default for active_storage.track_variants (true)
Has no impact on CCCD.

Enables tracking of uploaded file variants (if there are any) in the
database rather than by querying s3.

We do not generate variants therefore this setting has no impact. This
is an on-demand service provided out of the box by rails if needed.
Typically this is, for example, storing resized image version such
as thumbnails. We do, however, have the DB tables to do so if needed
and therefore this setting will not break existing functionality and
could be used if needed.