Load rails 6.1 default config using helper and promote non-defaults
Final step to adopting or not adopting default rails
config from rails 5.0 to rails 6.1.

Any config that is default according to the the result of
[config.load_defaults](https://guides.rubyonrails.org/configuring.html#results-of-config-load-defaults)
can be loaded using the rails helper `confg.load_defaults x.x`.

Any config that is non-default needs to be explicitly specified
in `config/application.rb` in order to overide those defaults.