Extend `error_detail_collection`
Add `long_messages_for` to `ErrorDetailCollection` for use
in error_presenter. I have added an `api_messages_for` for
completeness since the custom error handling has the 3 types
of error message (long, short and API).

The aim is to allow us to inject a presenter into to `govuk-formbuilder`
gems `govuk_error_summary` helper and have that use the
`long_messages_for`, or an alias, to display in the
summary instead of its default activerecord based error
handling.