Skip writing errors to database
Before stats reports started being written to AWS they were being stored in the
`report` field of the `stats_reports` database table. Now this field is mainly
unused but the error handling is still writing the error to this field. This
commit stops doing this as the `report` field is not being be accessed other
than by directly querying the database.