Change translation messages to work for both summary and field errors
It has not been possible to get the form builder working for
field level errors on nested objects based on our custom validator,
even when injecting a presenter.

This is because the presenter uses a very specific lookup for nested
model object errors that requires the "attribute" to have the nesting
detail in it, for example `foos_attribues_1_baz`. The form builder only
has access to the actual attribute name, `baz` in our example.

This alternative relies on two facts:

1. rails validation errors FOR FIELD ERRORS will be stored in conventional
translation locations location - https://guides.rubyonrails.org/i18n.html#error-message-scopes

2. govuk-formbuilder retrieves error messages set on the attribute and
displays them on the field level for any fields, nested or otherwise.

The translator can then be used for summary and api errors only. field
errors must be stored in the conventional location.

The translator converts each error message on an object, underscoring and
downcasing it, before it the looks up that error "key" on the attribute in
translations.

So an activemodel error of:
```
activerecord:
  errors:
    models:
      provider:
        attributes:
          role:
            blank: 'Choose at least one role'
```

becomes a tranlsator lookup in custom `provider.yml` translations of:
```
role:
  choose_at_least_one_role:
    long: blah
    short: _cannot_be_looked_up
    api: blah
```

TODO:
 - move convential model error translations to there own model files