Return filename after successful document create
The `document_file_name` field has been removed from the document model. The
whole model was being returned after a new document had been successfully
uploaded and this field was used to give the user feedback. As only this and
the document id is actually required the return data has been replaced with
a much smaller payload, taking the filename from the Active Storage attachment.