Implement `formatted_error_messages` method
This is the method expected by the govuk-formbuilder
`govuk_error_summary` helper for any injected presenter.

To behave as the current custom summary error presenter
it needs to add one summary line item per error per attribute.

To achieve this we need to construct an array
or attribute-error pairs which includes a
repeated attribute per message for that attribute.