Move initialize and call into Claim::Cleaner
The initialize and call methods are common to all the cleaners so can be moved
into the parent class. To allow for this the various methods for destroying
invalid fees are renamed with the generic name `destroy_invalid_fees`.

The cleaners will now each do the following three actions;

* destroy invalid fees
* destroy invalid disbursements
* clearn inapplicable fields

The default action in each case is to do nothing but either method is replaced
in the child class as required.