Fix non-rendering integers in `case_details/summary` view
The passing of a curly brace block with this component
helper does not work for non-string objects.

This is due to the block being used by a capture statement.
The implementation of [capture](https://github.com/rails/rails/blob/83217025a171593547d1268651b446d3533e2019/actionview/lib/action_view/helpers/capture_helper.rb#L46)
means it will only output the result if it is a `String` object.

This [commit](https://github.com/ministryofjustice/Claim-for-Crown-Court-Defence/commit/036af87b09b72f574d3173bc2fae8b75c3cb6def) introduced
the `govuk_summary_list_row` component and replaced use of, for example,
```ruby
.govuk-summary-list__row
  %dt.govuk-summary-list__key
    = t('estimated_trial_length', scope: trial_detail_fields)
  %dd.govuk-summary-list__value
    = claim&.estimated_trial_length
```
with
```ruby
= govuk_summary_list_row( t('actual_trial_length', scope: trial_detail_fields) ) { claim&.actual_trial_length }
```
...but since claim&.actual_trial_length is an integer it is not "captured" by the helper.

hamls '=' method would return a string so the code below would work
```ruby
= govuk_summary_list_row( t('actual_trial_length', scope: trial_detail_fields) ) do
  = claim&.actual_trial_length
```

if a single line with curly braces block is to be used then the object must be explictly cast to a string, as below
```ruby
= govuk_summary_list_row( t('actual_trial_length', scope: trial_detail_fields) ) { claim&.actual_trial_length&.to_s }
```

Instead of having to cast non-strings to strings when using the helper
we can enable plain blocks to be passed to `govuk_summary_list_row`
 inline `{ my_attribute.value }` and cast to string in the helper itself.

This may be generally advisable for all component helpers to
avoid silent failures.

In the end we need to either have view or feature specs to define the expected
view result so such refactorings do not silently break the views expected behaviour.