Move capybara extensions and extend
Move to a more convention `lib` like structure
so it can be shared more clearly with both
cucumber feature tests and rspec specs.

Will be using capybara to write view specs
for the claims summary page as these have a
lot of conditional and indent sensitive logic
that can easily regress.