Use subject caseworker name if imcomplete journey
If the claim is 'allocated', meaning that the final journey is incomplete,
the MI report has the name of the caseworker of the subject of the latest
transition but the author of the final transition for all previous journeys.

Formerly, the caseworker of the claim was used in the MI report for all
journeys when the claim was in an 'allocated' state.