Convert postgres timestamp to BST/GMT
Postgres stores dates as UTC so we need to handle
timzone changes explcitly.

UTC is inline with Europe/London when GMT is active but not
when it is inactive i.e. when UK siwtches to british summer
time. BST is 1 hour ahead of UTC/GMT.

```
BST begins at 01:00 GMT on the last Sunday of March and
ends at 01:00 GMT (02:00 BST) on the last Sunday of October.
```

Rails handles this implicitly but postgres does not so we
need to amend it using db helpers.