Ensure postgres original_submission_date dates retrieved as local timezone
Because we are retrieving dates using raw SQL and because
postgres stores dates in UTC (postgres default, and rails
default, albeit configurable) we need to explicitly convert
to Europe/London to handle daylight saving.

This impacted the diff because datetimes saved to DB during
DST/BST between 00:00:01 and 00:59:59 are saved as UTC (an
hour earlier) and when retrieved without conversion back to
BST are therefore a day earlier.

NOTES: Also, need to handle this boundary for 6 month retrieval