Sync claim_total rounding and kernal format rounding between v1 and v2
3+ decimal digits can cause minor differences. `Kernel#format` to 2
decimals rounds 3 decimals of 5 down while `my_decimal.round(2)` will
round it up.

To avoid the problem and build in monetary like precision it is recommended
to:
a. round the big decimal to 4 decimal places
b. then use Kernel#format

This should provide the required (perhaps OTT) precision and consistency
between the reports.