Try optimization using `find` to identify matching journey
OPTIMIZE: there should be only one submission, allocation or completion state per
journey as each journey slice starts with a submission, has all but one allocation
removed and ends in a completion state.

Note, however it is possible to not start with a submission type or allocation
if the that state transition was over 6 months ago.