Handle edge case of "stuck" and "unstuck" claims for `allocated_at` field
A long lived bug, possibly resulting from network issues, can
result in claims that have been alocated but then submitted again
and allocated again - referred to as "stuck" claims.

This results in claims with a journey of [submitteed, allocated, submitted, allocated, ...].
In such cases we have to find the last allocated date for the `allocated_at` data.

NOTE: The state machine allows an allocated claim to be submitted
possibly because it needs to revert a deallocated claim to its original
submitted state but not sure.