Add beta tag to new versions of MI report
For clarity and emphasis to the user these have had
a GDS styled tag.

Because the translations are all called based on the
report name I have had to suffix them all `_html` in order
to add html to just these the beta versions - This can be
undone as and when the new version replaces the old versions.